Move imperative files back into main package, for now at least
